% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_decode.R
\name{ais_decode_filter}
\alias{ais_decode_filter}
\alias{ais_decode_list}
\title{Decode NMEA Messages}
\usage{
ais_decode_filter(
  msgs,
  msg_type = ais_msgs$msg_1_2_3,
  data_type = c("auto", "file", "strings", "text"),
  as_sf = maritime_as_sf(),
  as_tibble = maritime_as_tibble(),
  verbose = maritime_verbose()
)

ais_decode_list(
  msgs,
  data_type = c("auto", "file", "strings", "text"),
  as_sf = maritime_as_sf(),
  as_tibble = maritime_as_tibble(),
  verbose = maritime_verbose()
)
}
\arguments{
\item{msgs}{\code{character()}
\itemize{
\item NMEA messages
}}

\item{msg_type}{See \code{ais_msgs}}

\item{data_type}{\code{character(1L)}
\itemize{
\item how to interpret \code{x}.
\itemize{
\item one of \code{"auto"}, \code{"file"}, \code{"strings"}, or \code{"text"}
\itemize{
\item \code{"auto"}: let \code{ais_decode()} decide
\item \code{"file"}: treat \code{x} as a file path
\item \code{"strings"}: treat \code{x} as a vector with each element containing a single AIS message
\item \code{"text"}: treat \code{x} as a single \code{character} containing multiple AIS messages
}
}
}}

\item{as_sf}{\code{logical(1L)}, default: \code{maritime_as_sf()}
\itemize{
\item Whether an \code{sf} (\code{\link[sf:st_sf]{sf::st_sf()}}) should be returned.
\item Ignored if the \code{{sf}} package is not installed.
\item \code{maritime_as_sf()} defaults to \code{TRUE}.
}}

\item{as_tibble}{\code{logical(1L)}, default: \code{maritime_as_tibble()}
\itemize{
\item Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned.
\item Ignored if the \code{{tibble}} package is not installed.
\item \code{maritime_as_tibble()} defaults to \code{TRUE}.
}}

\item{verbose}{\code{logical(1L)}, default: \code{maritime_verbose()}
\itemize{
\item Whether print status and diagnostic messages.
\item \code{maritime_verbose()} defaults to \code{TRUE}.
}}
}
\description{
National Marine Electronics Association (NMEA) messages are comma separated value
lines of text that start with sender or "talker" code and the sentence type.
}
\details{
\code{ais_decode_filter} decodes NMEA messages (currently 1, 2, and 3) contained in files or \code{character} vectors.
}
\examples{
nmea_file <- example_nmea_file()
nmea_lines <- readLines(nmea_file, warn = FALSE)

nmea_lines[1:10]

# decoding a file, filter by messgage type ============================================
ais_decode_filter(nmea_file, msg_type = ais_msgs$msg_1_2_3)
ais_decode_filter(nmea_file, msg_type = ais_msgs$msg_4_11)
ais_decode_filter(nmea_file, msg_type = ais_msgs$msg_5)

# decoding a file, keeping everything (WIP) ===========================================
ais_decode_list(nmea_file)

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
