% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ais_version}
\alias{ais_version}
\title{\code{libais} Version Information}
\usage{
ais_version()
}
\description{
Version info for \href{https://github.com/schwehr/libais}{libais}.
}
\examples{
libais_version()

}
