% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_example_data.R
\name{example_nmea_file}
\alias{example_nmea_file}
\title{Path to Example File}
\usage{
example_nmea_file()
}
\value{
\code{character(1L)} file path
}
\description{
Path to a file containing example NMEA messages.
}
\examples{
example_nmea_file()

}
\author{
Christopher Callaghan \email{cjcallag@nps.edu}
}
