% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maritime_options.R
\name{maritime_options}
\alias{maritime_options}
\alias{maritime_parse}
\alias{maritime_as_df}
\alias{maritime_as_sf}
\alias{maritime_as_tibble}
\alias{maritime_add_query_cols}
\alias{maritime_query_col_prefix}
\alias{maritime_snakecase_col_names}
\alias{maritime_return_raw_response}
\alias{maritime_verbose}
\title{\code{{maritime}} Options}
\usage{
maritime_options()

maritime_parse()

maritime_as_df()

maritime_as_sf()

maritime_as_tibble()

maritime_add_query_cols()

maritime_query_col_prefix()

maritime_snakecase_col_names()

maritime_return_raw_response()

maritime_verbose()
}
\description{
\code{{maritime}} Options
}
\examples{
maritime_options()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
